requirements.txt
fastapi
uvicorn
aiohttp
python-dotenv
pydantic
numpy